@@ -1,61 +1,54 @@
 {
-  "name": "spatial-navigation-like-youtubetv-app",
+  "name": "spatial-navigation-site",
+  "version": "1.0.0",
+  "main": "./bin/server/index.js",
   "scripts": {
-    "dev": "node ./bin/server",
     "clean": "rimraf ./build",
-    "build": "npm run clean && cross-env DEBUG=app:* NODE_ENV=development CONFIG=development node ./bin/compile",
-    "build:prod": "npm run clean && cross-env DEBUG=app:* NODE_ENV=production CONFIG=production node ./bin/compile",
-    "start:prod": "cross-env NODE_ENV=production CONFIG=production node ./bin/server"
+    "dev": "nodemon",
+    "build:dev": "npm run clean && cross-env DEBUG=app:* NODE_ENV=development node bin/compile",
+    "build:prod": "npm run clean && cross-env DEBUG=app:* NODE_ENV=production node bin/compile"
   },
   "dependencies": {
     "classnames": "^2.2.5",
-    "hoist-non-react-statics": "^1.2.0",
-    "normalize.css": "^4.2.0",
-    "react": "^15.2.1",
-    "react-css-modules": "^3.7.7",
-    "react-custom-scrollbars": "^4.0.0",
-    "react-dom": "^15.3.1",
+    "normalize.styl": "^3.0.3",
+    "react": "^15.3.2",
+    "react-css-modules": "^3.7.10",
+    "react-dom": "^15.3.2",
     "react-fa": "^4.1.2",
-    "react-router": "^2.5.2",
-    "reflexbox": "^1.1.4"
+    "react-highlight.js": "^1.0.5",
+    "react-router": "^3.0.0",
+    "reflexbox": "^2.2.2"
   },
   "devDependencies": {
-    "app-module-path": "^1.1.0",
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.0.20",
-    "babel-loader": "^6.0.1",
+    "app-module-path": "^2.1.0",
+    "babel-core": "^6.17.0",
+    "babel-loader": "^6.2.5",
     "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-react-transform": "^2.0.2",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-runtime": "^6.12.0",
-    "babel-preset-es2015": "^6.0.15",
-    "babel-preset-react": "^6.0.15",
-    "babel-preset-stage-0": "^6.0.15",
-    "babel-register": "^6.9.0",
+    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-react": "^6.16.0",
+    "babel-preset-stage-0": "^6.16.0",
+    "babel-register": "^6.16.3",
     "body-parser": "^1.15.2",
     "chalk": "^1.1.3",
-    "cross-env": "^2.0.0",
-    "css-loader": "^0.23.1",
+    "connect-history-api-fallback": "^1.3.0",
+    "cross-env": "^3.1.3",
+    "css-loader": "^0.25.0",
     "debug": "^2.2.0",
     "express": "^4.14.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
-    "glob": "^7.1.0",
-    "handlebars": "^4.0.5",
-    "handlebars-loader": "^1.3.0",
-    "html-webpack-plugin": "^2.22.0",
-    "nib": "^1.1.0",
-    "react-hot-loader": "^1.3.0",
-    "require-dir": "^0.3.0",
-    "resolve-url-loader": "^1.6.0",
-    "rimraf": "^2.5.3",
+    "html-webpack-plugin": "^2.24.0",
+    "nib": "^1.1.2",
+    "nodemon": "^1.11.0",
+    "require-dir": "^0.3.1",
+    "rimraf": "^2.5.4",
     "style-loader": "^0.13.1",
     "stylus": "^0.54.5",
-    "stylus-loader": "^2.1.1",
+    "stylus-loader": "^2.3.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.12.2",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-hot-middleware": "^2.12.1"
+    "webpack": "^1.13.2",
+    "webpack-dev-middleware": "^1.8.4",
+    "webpack-hot-middleware": "^2.13.0"
   }
 }
