@@ -10,11 +10,13 @@ debug(`Environment is set to: ${process.env.NODE_ENV || 'default'}`)
 debug('Webpack compiler starting to build')
 
 
-webpack(webpackCfg).run((err, stats) => {
+const compiler = webpack(webpackCfg)
+
+compiler.run((err, stats) => {
   const jsonStats = stats.toJson()
 
   debug('Compilation completed!')
-
+  
   console.log(stats.toString({
     colors: true,
     children: false,
@@ -24,14 +26,15 @@ webpack(webpackCfg).run((err, stats) => {
   if (err) {
     debug(chalk.red(err))
     process.exit(1)
-  }
+  } 
   else if (jsonStats.errors.length > 0) {
     debug(chalk.red(jsonStats.errors))
     process.exit(1)
-  }
+  } 
   else if (jsonStats.warnings.length > 0) {
     // debug(chalk.yellow(jsonStats.warnings))
   }
 
   debug('All done - everything is good to go.')
+  process.exit(0)
 })
