@@ -5,7 +5,5 @@ import Home from 'containers/Home'
 
 
 export default (
-  <Route>
-    <Route path="/" component={Home} />
-  </Route>
+  <Route path="/" component={Home} />
 )
