@@ -1,7 +1,13 @@
+import config from '../config'
+
+
 export default (webpackConfig) => {
-  webpackConfig.entry.push(
-    'webpack-hot-middleware/client?path=/__webpack_hmr'
-  )
+  webpackConfig.devServer = {
+    publicPath: webpackConfig.output.publicPath,
+    stats: 'errors-only',
+    noInfo: true,
+    lazy: false,
+  }
 
   return webpackConfig
 }
